Move Trigger into its own file.
