+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/AbstractTaggedTriggerManager.java
@@ -1,7 +1,6 @@
 package io.github.wysohn.triggerreactor.core.manager.trigger;
 
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
-import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager.Trigger;
 
 import java.io.File;
