+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/named/AbstractNamedTriggerManager.java
@@ -18,6 +18,7 @@ package io.github.wysohn.triggerreactor.core.manager.trigger.named;
 
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
 import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager;
+import io.github.wysohn.triggerreactor.core.manager.trigger.Trigger;
 import io.github.wysohn.triggerreactor.core.script.lexer.LexerException;
 import io.github.wysohn.triggerreactor.core.script.parser.ParserException;
 import io.github.wysohn.triggerreactor.tools.FileUtil;
