+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/named/NamedTrigger.java
@@ -1,10 +1,11 @@
 package io.github.wysohn.triggerreactor.core.manager.trigger.named;
 
 import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager;
+import io.github.wysohn.triggerreactor.core.manager.trigger.Trigger;
 
 import java.io.File;
 
-class NamedTrigger extends AbstractTriggerManager.Trigger {
+public class NamedTrigger extends Trigger {
 
     public NamedTrigger(String name, File file, String script) throws AbstractTriggerManager.TriggerInitFailedException {
         super(name, file, script);
@@ -13,7 +14,7 @@ class NamedTrigger extends AbstractTriggerManager.Trigger {
     }
 
     @Override
-    public AbstractTriggerManager.Trigger clone() {
+    public Trigger clone() {
         try {
             return new NamedTrigger(triggerName, file, getScript());
         } catch (AbstractTriggerManager.TriggerInitFailedException e) {
